requirements.txt
Django>=2.0,<3.0
dj-database-url==0.5.0
django-cors-headers==3.1.0
django-filter==2.3.0
djangorestframework==3.10.2
gunicorn==19.9.0
Markdown==3.1.1
psycopg2>=2.7,<3.0
python-decouple==3.1
whitenoise==4.1.2
